image_gen/requirements.txt
requests
pymongo
python-dotenv
Pillow
fastapi
uvicorn
